10.31 Packages that cannot or should not be built  =================================================
There are several reasons why a package might be instructed to not
build under certain circumstances. If the package builds and runs
on most platforms, the exceptions should be noted with NOT_FOR_PLATFORM.
If the package builds and runs on a small handful of platforms,
set ONLY_FOR_PLATFORM instead. If the package should be skipped
(for example, because it provides functionality already provided
by the system), set PKG_SKIP_REASON to a descriptive message. If
the package should fail because some preconditions are not met,
set PKG_FAIL_REASON to a descriptive message.

IGNORE is deprecated because it didn't provide enough information
to determine whether the build should fail.