Initial import of dvipdfmx-20021230 to the NetBSD packages collection. It depends on newly imported teTeX2 package.
The dvipdfmx (formerly dvipdfm-cjk) project provides an eXtended version
of the dvipdfm, a DVI to PDF translator developed by Mark A. Wicks.  The
primary goal of this project is to support multi-byte character
encodings and large character sets for East Asian languages by CID-keyed
font technology.  The secondary goal is to support as many features as
pdfTeX developed by Han The Thanh.  This project is a combined work of
the dvipdfm-jpn project by Shunsaku Hirata and its modified one,
dvipdfm-kor, by Jin-Hwan Cho.

It has many other features.  For further info, visit the projects' web
site:

    http://project.ktug.or.kr/dvipdfmx/