Add patch from Todd Vierling in PR 21226: The following method overrides "libgcc_link_spec" in the generated specs file, which gives you a compiler that automatically inserts the correct rpath argument at link time.  This should work for any pkgsrc target platform, as no platform-specific config overrides it.
Additionally, part of this fix changes the "scrubbed" LDFLAGS by
overriding that in CONFIGURE_ENV, rather than globally.  Other changes to
the pkgsrc Makefile (including one I plan to submit for threads support)
may need access to the "real" LDFLAGS, so overriding it wholesale in the
pkg Makefile is Bad.

Bump version to 3.2.2nb2.