<< Always test a package on sparc64 before committing! >> ;)
As seen in libc, prevent gcc's memory usage from exploding during
optimizing of SHA code by moving each round into a separate function on
sparc64.

Also remove TEXINFO_REQ, scli.{texinfo,info} are patched and touched in
such a way that makeinfo is not required for build and of course it was
not for "running": devel/gtexinfo is no more a dependency on most
NetBSD system and most of !NetBSD pkgsrc platforms.

XXX This TeXinfo/Info files issue will be more cleanly addressed soon
and wholesale...

Bump PKGREVISION.