Import of NetBSD-current's pax (replacing the older pkgtools/pax).
NetBSD pax is the POSIX standard archiver with many extensions.

pax will read, write, and list the members of an archive file, and will
copy directory hierarchies.

pax operation is independent of the specific archive format, and supports
a wide variety of different archive formats.

This package uses pkgtools/libnbcompat, and is known to build on
Solaris 9, Slackware 8.1 and Darwin 6.6. Please send-pr(1) if this
package is broken on your platform.