Initial import of beecrypt, version 2.1.0:
BeeCrypt is an open source cryptography library that contains highly
optimized C and assembler implementations of many well-known algorithms
including Blowfish, MD5, SHA-1, Diffie-Hellman, and ElGamal.

Unlike some other crypto libraries, BeeCrypt is not designed to solve
one specific problem, like file encryption, but to be a general purpose
toolkit which can be used in a variety of applications.

There are also no patent or royalty issues associated with BeeCrypt, and
it is released under the GNU LGPL license, which means it can used for
free in both open source and closed source commercial projects.