Import p5-Scalar-List-Utils version 1.12.
This distribution is a replacement for the builtin distribution.

This package contains a selection of subroutines that people have
expressed would be nice to have in the perl core, but the usage would not
really be high enough to warrant the use of a keyword, and the size so
small such that being individual extensions would be wasteful.