Update to gwave-20030804.
This update is a collection of little improvements and a significant
verhaul of hardcopy plot and export.

Hardcopy and documentation output can now be obtained using either graph
or Gnuplot.  A dialog box allows choosing output formas and a few other
settings.  Additional plot/export filters can be plugged in by writing a
suitable scheme module.

There's a data-export function for extracting and saving a subset of a
datafile as ascii.

The stacking/drawing order of waveforms can be changed.

Selected waveforms are now drawn in a highlighted style.