Initial import of profit, one of the many packages provided by brook at biology dot nmsu dot edu and his team at NMSU.
ProFit (pronounced Pro-Fit, not profit!) is designed to be the ultimate
program for performing least squares fits of two protein structures. It
performs a very simple and basic function, but allows as much flexibility as
possible in performing this procedure. Thus one can specify subsets of atoms
to be considered, specify zones to be fitted by number, sequence, or by
sequence alignment.