Update to version 4.08.
Changes since 4.07:
* fix `sed n' printing the last line twice.
* fix incorrect error message for invalid character classes.
* fix segmentation violation with repeated empty subexpressions.
* fix incorrect parsing of ^ after escaped (.
* more comprehensive test suite (and with many expected failures...)

Please note that since NetBSD's sed is available in a portable
version in pkgsrc as textproc/nbsed, and in bootstrap-pkgsrc I do
not consider this package as being low in the dependencies food-chain
anymore. This is especially true for platforms other than NetBSD.

So let me explicitly add dependencies on libiconv and gettext-lib packages
(if needed of course).