Initial import of blccc-1.99: Blinkenlights Chaos Control Center
This is a rewrite of the master control program for Blinkenlights.
The first implementation used to be a GTK+ application that allowed
previewing of movies and run-time modification of the playlist using
a graphical user interface.  The need for a way to control
the Blinkenlights remotely arose, so this new implementation runs as
a daemon controllable via telnet connections.  This software was used
for the Blinkenlights installation at the end of September 2001.


Greetings from 20c3, 20th Chaos Communication Congress in Berlin, where the
real Blinkenlights are running at this very moment on the Haus des Lehrers,
next to the congress building ;) (http://www.blinkenlights.de/).