Initial import of icecast2-2.0.0, packaged by pancake in pkgsrc-wip, adapted to use the bsd.pkg.install.mk framework by me.
Icecast is a streaming media server which currently supports Ogg
Vorbis and MP3 audio streams. It can be used to create an Internet
radio station or a privately running jukebox and many things in
between. It is very versatile in that new formats can be added
relatively easily and supports open standards for commuincation and
interaction.

--
The package has been prepared to be launched in chroot mode by default,
enjoy!