Initial import of honeyd-0.8b, from pkgsrc-wip packaged by Adrian Portelli with minor changes by me.
Honeyd is a small daemon that creates virtual hosts on a network. The hosts
can be configured to run arbitrary services, and their personality can be
adapted so that they appear to be running certain operating systems. Honeyd
enables a single host to claim multiple addresses - I have tested up to
65536 - on a LAN for network simulation. Honeyd improves cyber security by
providing mechanisms for threat detection and assessment. It also deters
adversaries by hiding real systems in the middle of virtual systems.