Import p5-Locale-Maketext 1.08
It is a common feature of applications (whether run directly, or via
the Web) for them to be "localized" -- i.e., for them to a present an
English interface to an English-speaker, a German interface to a
German-speaker, and so on for all languages it's programmed with.
Locale::Maketext is a framework for software localization; it provides
you with the tools for organizing and accessing the bits of text and
text-processing code that you need for producing localized applications.