Update misc/uptimec to 0.2p3, closes PR pkg/26082.
Changes:

Version 0.2p3
-------------
Fixed problem when compiling on some HP/UX systems.
Implemented changes that forces uptimec to use select() on HP/UX
systems even if sys/select.h is not availible.


Version 0.2p2
-------------
Imported new routines to extract uptime on Tru64 which enables the client
to run in a chrooted environment.
Will now bind() socket before using it, which seems to be required on some
platforms. (Noted when running in FreeBSD jail)

Contributors to this release:
        Tobias Nygren
        Aaron D. Gifford