Initial import of aranym-0.8.8beta.
ARAnyM is an acronym: Atari Running on Any Machine. It is a virtual machine
software for running the Atari ST/TT/Falcon operating systems (TOS, FreeMiNT,
MagiC and others) and TOS/GEM applications on any kind of hardware - be it
an IBM clone (read it as "PC" :-), an Apple, an Unix server, a graphics
workstation or even a portable computer.