There is no way for libtool to generate a convenience library that can be used to create shared libraries _and_ be linked into a statically linked program.  Instead of trying to hack libtool to do this, just accept the fact and do what you want another way!
Remove the ugly hack in sasldb/Makefile.am that regenerated the static
archive from the non-PIC object files.  While this was fine for linking
into programs, it breaks things when you link this into the sasldb
plugin.  Leaving it the other way, where the static archive from the
PIC object files, is also unacceptable because there are potential
problems on some platforms when linking an archive of of PIC objects
into an executable.  The solution: let the static archive contain PIC
objects and be used to link into the sasldb plugin, but for the programs
in the utils directory, explicitly add the non-PIC object files listed
in $(SASL_DB_BACKEND_STATIC) to the files used to generate the programs.
This is easy because SASL_DB_BACKEND_STATIC is already generated
properly by config/sasldb.m4, so make use if it.

This should fix problems with using Cyrus SASL on non-i386 platforms
noted in PR pkg/26492 by Matt Dainty.  Bump the PKGREVISION.