Re-import gtl as gtl0, since this (older) version is needed for math/cassowary and I'll update gtl in a minute.
DESCR:
GTL is a graph library based on STL.  For the design of GTL's API the
API of LEDA has served as a basis.  GTL contains the classes needed to
work with graphs, nodes and edges and some basic algorithms (DFS, BFS,
topsort ...) as building blocks for more complex graph algorithms.

This package contains an older version of the library, for a newer
version use the gtl package.