Initial import of pflkm-20041025 into the NetBSD Packages Collection.
Packet Filter (from here on referred to as PF) is OpenBSD's system for
filtering TCP/IP traffic and doing Network Address Translation. PF is also
capable of normalizing and conditioning TCP/IP traffic.

PF was originally developed by Daniel Hartmeier and is now maintained and
developed by Daniel and the rest of the OpenBSD team.

This package includes a complete port (LKM and userland utilities) from
OpenBSD 3.6 to NetBSD 2.0.