Initial import of guile-lib, provided by Greg Troxel via pkgsrc-wip.
Guile-Lib is intended as an accumulation place for pure-scheme Guile
modules, allowing for people to cooperate integrating their generic
Guile modules into a coherent library. Think "a down-scaled,
limited-scope CPAN for Guile".

Also, it can be seen as a code staging area for Guile; the Guile
developers could decide to integrate some of the code into
guile-core. An example for a possible candidate is SRFI-35.

    * SSAX, S-Exp-based XML parsing/query/conversion
    * HTMLPrag, a permissive ("pragmatic") HTML parser
    * Texinfo processing, enabling literate programming
    * Unit testing framework ala JUnit
    * Logging system
    * String routines (wrapping, completion, soundex algorithm)
    * OS process chains (think "shell pipes in scheme")
    * An LALR parser
    * ANSI escape sequence text coloring
    * Structured text (plain, texinfo, html)
    * SRFI-35 (conditions)