Import p5-Spoon-0.21 as devel/p5-Spoon, a Spiffy Application Building Framework.
Spoon is an Application Framework that is designed primarily for
building Social Software web applications.  Spoon::Base is the primary
base class for all the Spoon::* modules.  Spoon.pm inherits from
Spiffy.pm.  Spoon is not an application in and of itself.  You need
to build your own applications from it.