Update to 1.6.7:
Changes since Guile 1.6.6 (changes in 1.6.7):

* Changes to the distribution

** A build problem has been fixed.

Previously, on some systems, the build would fail when libguile-ltdl
couldn't be found during the build.  This should now be fixed.

* Changes to Scheme functions and syntax

** array-map! and array-map-in-order! now correctly require at least one source

A mistake caused a call with just one source array to be rejected,
this has been fixed.

** string->number and octal constant bignums

An incorrect bignum size calculation has been fixed, this caused
overflow errors in string->number on bases other than 2, 10 and 16,
including octal literal constants in code or the reader.

** SRFI-1 alist-delete equality argument order fixed.

In the srfi-1 module alist-delete and alist-delete!, the order of the
arguments to the "=" procedure now matches the SRFI-1 specification.

** SRFI-13 string-any and string-every tail calls

string-any and string-every now make a tail call to their predicate
function on reaching the last character in the string, per the SRFI-13
specification.