Initial import of kajaani-kombat:
Kajaani Kombat is a funny multiplayer game... and much more! It is
a Rampart-like game (old arcade classic) set in space.

Every player has a castle demarked by walls, and cannons to shoot
at other player's castles. After the shooting phase there is a
repair phase where one has to rebuild the walls such that one there
is an area completely encased by walls (without holes). If one
doesn't succeed with this, one loses.

Kajaani Kombat is playable with two to four players, over the
internet or alternatively two players sharing one computer. It is
also possible to play over the internet with, for example, 4 players
of which two are sharing the same computer. Enjoy it with your
friends!