Add PLIST.FreeBSD and three patches to make it build for FreeBSD.
Add patch-ak and patch-an to support -enable-fpic with recent
versions of gcc.  From xine cvs.

Bump PKGREVISION just in case these patch-ak and patch-an change
functionality.

Tested builds and running on FreeBSD 5.4-STABLE with gcc 3.4.2 and
FreeBSD 5.3, and Linux 2.6.9 with gcc 3.3.5. And tested build
and install on NetBSD 1.6.2_STABLE.

Okayed by jmmv a couple weeks ago.