Initial import of p5-File-Flat-0.95:
File::Flat implements a flat file system. A flat file system is a
file system in which directories do not exist. It provides an
abstraction over any normal file system which makes it appear as if
directories do not exist. In effect, it will automatically create
directories as needed. This is create for things like install
scripts and such, as you never need to worry about the existence
of directories, just write to a file, no matter where it is.