Alter the configuration so that:
	$(prefix) == $(siteprefix) == $(vendorprefix)

so that if a perl module is configured with "perl Makefile.PL PREFIX=...",
then that single PREFIX definition will override all three of the
above, and files will be properly installed into the correct relative
path.  Also, patch a test so that it understands the different behavior
of the pkgsrc ExtUtils::MakeMaker module.  Bump the PKGREVISION to 2.