Include sys/types.h. This fixes configure on DragonFly. Bump PKGREVISION.
Okayed by lha@. I tested on Linux and DragonFly. I got this from
Joerg Sonnenberger.

On DragonFly, the configure errored like:

/usr/include/openssl/md5.h:110: error: syntax error before "size_t"
In file included from conftest.c:34:
/usr/include/openssl/sha.h:109: error: syntax error before "size_t"

This caused tests to break and it ended up building and installing libdes
and des.h, md4.h, and related headers.

So later libgssapi needed this libdes which was not buildlinked which
broke kdelibs3 build.