Changes 0.9.4:
Took some time to do this release, and the fixes are numerous, an
upgrade is highly recommended. Major changes include an increased
header limit, which fixes inter-operation with some versions of lynx;
a fix for a crash when trying to access user pages of people who do
not actually exist (ie, /~badusername).

On the new features front, the -U option now accepts usernames too,
not just user ids.