Initial import of nagios-base 2.0, based on the 1.3.x package in pkgsrc-wip.
Nagios is a program that will monitor hosts and services on your
network.  It has the ability to email or page you when a problem arises
and when it gets resolved.  Nagios is written in C and is designed to
run under Linux, although it should work under most other *NIX variants.
It can run either as a normal process or as a daemon, intermittently
running checks on various services that you specify.  The actual service
checks are performed by external "plugins" which return service
information to Nagios.  Several CGI programs are included with Nagios
in order to allow you to view the current service status, history, etc.
via a web browser.

This package provides the core nagios infrastructure and web pages.