Pioneers is a clone of the board game The Settlers of Catan. Pioneers was formerly known as Gnocatan.
This is an 'update' based on the old gnocatan package with some changes:
- some dependencies changes, so it needs fewer packages
- fix hardcoded manpage paths

There are too many changes since gnocatan 0.8.1.16 to list them here.
Please check the NEWS file in the source archive for more information.

There is one important one:

CVS Snapshot release 0.9.55
This is a security release for the meta server.
* The public meta server could be crashed by sending commands to create
  a new game, even though the meta server does not allow the creation
  of new games.
 Everybody who is running a meta server is requested to upgrade.

http://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2006-0467