Initial addition of p5-SVN-Notify version 2.57 as devel/p5-SVN-Notify in the NetBSD Packages Collection.
The Perl5 module SVN::Notify provide a class that may be used for
sending email messages for Subversion repository activity. There
are a number of different modes supported, and SVN::Notify is fully
subclassable, to easily add new functionality. By default, A list
of all the files affected by the commit will be assembled and listed
in a single message. An additional option allows diffs to be
calculated for the changes and either appended to the message or
added as an attachment. The included subclass, SVN::Notify::HTML,
allows the messages to be sent in HTML format. A svnnotify script
is also provided.