Update to i2cbd-2.0_BETA5: fold in patch-aj, from Thor Lancelot Simon:
Don't set O_NDELAY (and leave it set!) on all sockets, all the time.  This
was making select always return immediately and was causing the daemon to
spin on CPU.

Instead, set O_NONBLOCK (= O_NDELAY) on sockets only when reading/writing
them, and clear it afterwards, so select is always safe.