Added computation of version number for package bzip2 if there is a native version available in the system.
The awk skript looks for the line in bzlib.h containing the version
information. This works for versions < 1.0.0. For versions >= 1.0.0 the
script additionally evaluates the copyright date.
According to the change history, version 1.0 and 1.0.1 are identical except
for a documentation update so we treat both as version 1.0.1 to avoid
changing BUILDLINK_API_DEPENDS.bzip2 in buildlink3.mk.